@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>net6.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <SignAssembly>False</SignAssembly>
     <Version>2.0.0</Version>
     <ServerGarbageCollection>true</ServerGarbageCollection>
