@@ -10,7 +10,7 @@ namespace BitFaster.Caching.Benchmarks
 #if Windows
     [SimpleJob(RuntimeMoniker.Net48)]
 #endif
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net90)]
     [MemoryDiagnoser(displayGenColumns: false)]
     public class DataStructureBenchmarks
     {
