@@ -12,7 +12,7 @@ namespace BitFaster.Caching.UnitTests
     public class CacheTests
     {
 // backcompat: remove conditional compile
-#if NETCOREAPP3_0_OR_GREATER
+#if NET
         [Fact]
         public void WhenCacheInterfaceDefaultGetOrAddFallback()
         {
