@@ -1,7 +1,5 @@
-﻿#if NETCOREAPP3_1_OR_GREATER
+﻿#if NET
 using System.Runtime.Intrinsics.X86;
-#endif
-#if NET6_0_OR_GREATER
 using System.Runtime.Intrinsics.Arm;
 #endif
 
@@ -13,15 +11,9 @@ namespace BitFaster.Caching.UnitTests
     {
         public static void SkipAvxIfNotSupported<I>()
         {
-#if NETCOREAPP3_1_OR_GREATER
-    #if NET6_0_OR_GREATER
+#if NET
             // when we are trying to test Avx2/Arm64, skip the test if it's not supported
             Skip.If(typeof(I) == typeof(DetectIsa) && !(Avx2.IsSupported || AdvSimd.Arm64.IsSupported));
-    #else
-            // when we are trying to test Avx2, skip the test if it's not supported
-            Skip.If(typeof(I) == typeof(DetectIsa) && !Avx2.IsSupported);
-    #endif
-
 #else
             Skip.If(true);
 #endif
