@@ -429,7 +429,7 @@ namespace BitFaster.Caching.UnitTests.Lru
         }
 
 // backcompat: remove conditional compile
-#if NETCOREAPP3_0_OR_GREATER
+#if NET
         [Fact]
         public void WhenKeyExistsTryUpdateIncrementsUpdateCount()
         {
