@@ -59,7 +59,7 @@ namespace BitFaster.Caching.UnitTests
         }
 
 // backcompat: remove conditional compile
-#if NETCOREAPP3_0_OR_GREATER
+#if NET
         [Fact]
         public void WhenUpdatedEventHandlerIsRegisteredItIsFired()
         {
@@ -83,7 +83,7 @@ namespace BitFaster.Caching.UnitTests
         }
 
         // backcompat: remove conditional compile
-#if NETCOREAPP3_0_OR_GREATER
+#if NET
         [Fact]
         public async Task WhenKeyDoesNotExistGetOrAddArgAddsValueWithArg()
         {
