@@ -1,6 +1,6 @@
 ﻿using System.Runtime.CompilerServices;
 
-#if NET6_0_OR_GREATER
+#if NET
 [module: System.Runtime.CompilerServices.SkipLocalsInit]
 #endif
 
