@@ -59,8 +59,8 @@ namespace BitFaster.Caching
         /// <returns>A task that represents the asynchronous ScopedGetOrAdd operation.</returns>
         ValueTask<Lifetime<V>> ScopedGetOrAddAsync(K key, Func<K, Task<Scoped<V>>> valueFactory);
 
-        // backcompat: remove conditional compile
-#if NETCOREAPP3_0_OR_GREATER
+// backcompat: remove conditional compile
+#if !NETSTANDARD
         /// <summary>
         /// Adds a key/scoped value pair to the cache if the key does not already exist. Returns a lifetime for either 
         /// the new value, or the existing value if the key already exists.
